Move TokenManager example to a new module
