+++ examples/src/lib.rs
@@ -4,3 +4,4 @@ pub mod mapping;
 pub mod ownable;
 pub mod owned_token;
 pub mod tlw;
+pub mod token_manager;
